ColorDisplayImpl.illuminateの排他処理の実装、コンパイルエラーのデバッグ

+++ Peer/Peer/ColorDisplay.swift
@@ -24,5 +24,5 @@ public protocol ColorDisplay
     ///   - interval: 色を変える間隔 [s]
     ///   - colors: 表示色（複数指定可能；要素順で表示）
     ///   - isRepeat: 最後の色まで到達したら先頭の色に戻って繰り返すか（true: 繰り返す、false: 繰り返さない）
-    func illuminate(interval: Int, colors: [Color], isRepeat: Bool)
+    func illuminate(interval: Double, colors: [Color], isRepeat: Bool)
 }
